requirements.txt
streamlit==1.31.0
tensorflow==2.15.0
pillow==10.2.0
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2
gdown==5.1.0
